Style changes after code review
