+++ backend/api/model/tweet.go
@@ -1,8 +1,6 @@
 package model
 
-import (
-	"time"
-)
+import "time"
 
 type Tweet struct {
 	ID           int64     `json: "author_id"`
@@ -16,18 +14,4 @@ type Tweet struct {
 type NewTweet struct {
 	AuthorID int64  `json: "id"`
 	Content  string `json: "content"`
-}
-
-type User struct {
-	ID        int64     `json: "id"`
-	Name      string    `json: "name"`
-	Username  string    `json: "username"`
-	Email     string    `json: "email"`
-	CreatedAt time.Time `json: "created_at"`
-}
-
-type NewUser struct {
-	Name     string `json: "name"`
-	Username string `json: "username"`
-	Email    string `json: "email"`
 }
