+++ backend/database/model/user.go
@@ -1,0 +1,11 @@
+package model
+
+import "time"
+
+type User struct {
+	ID        int64
+	Name      string
+	Username  string
+	Email     string
+	CreatedAt time.Time
+}
